userservice/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
jsonschema==3.2.0
packaging==25.0
paramiko==3.5.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
python-jose==3.5.0
PyYAML==5.3.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
texttable==1.7.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==0.59.0
